RHEL: output/output.ml: Remove cache=none
Virt-v2v tries to reduce the amount of page cache used when writing
out the disk image to the target storage. It does this by enabling the
cache=none option in nbdkit-file-plugin.

However, use of nbdkit-file-plugin cache=none option causes data
corruption with Dell PowerMax 8000 storage (only).

As this is only an advisory setting, remove it.  Use of the page cache
can be limited instead using cgroupsv2 memory settings.

Fixes: https://issues.redhat.com/browse/RHEL-135617